database done. Need to make backend work, at least getAll and getById. Just to see it working and finally get back to frontend
"Tried to edit the downloaded SQL file from MTGJson.com but I couldn´t do it for now, so I decided to go with it: with exceeding tables and such. So I build the DB, edited my personal SQL and ran on it. Now it has all the tables I need and the ones I don't. It's completely full of cards. So I went to test backend and... it's not ok. GetAll seems to work fine but it's too much data for POSTMAN (browser seems to run it, but laggy). GetById returned empty square brackets, so I'll have to figure this out next sess. Also, had to put the main SQL in a new .gitignore in the Project folder for instance, because it`s large."