some css improvement, and  some improvement in controllers.
Long day trying to figure out the perfect way to use search engine with the already implemented input options. In the end after dozens of hours, we're back to step one but with a few improvement like when typing and all colors selected it will show all cards (including multicolored and colorless). But I think I'll revert this to trigger when not typing, as it should be natural to see all your cards right by default. Though if you type, shouldn't it look for all cards with typed name? I'll try it next commit