overlay is working!
Count (onCollection) is working in each card. This means that on every page a card is rendered (collection, all, wishlist, decks), every card will display how many of it the user has and how many he wants on his wishlist.

Funny story: I was trying to implement a new hover effect - that insted of scaling the card itself, clones the card right next to it, scaled; and the scaled card moves with the cursor. Trying to implement it, overlay finally worked. It turns out that the .CardOverlay class had to be position:absolute; and .CardContainer AND .Card had to be position:relative; The image inside the Card Container had to be position:relative; as well.

I also changed overlay appearance (it's still provisory) and granted to it some animation for it to fade in at the same time as the card scales with border.

SearchContainer has been given more margin-bottom to not stay behind the first row of cards when they are hovered.