all sets implemented in search filter. It's fetching from backend and stored locally on browser.
note that there is localSetsVersion item in browser. Whenever it's different from frontend, it will re-fetch to make it so. So whenever there are db updates to include new sets, this must be changed in frontend too.

Search Container css improvements.

Deck Container div rendering improvements > now rendered only when manaValueArrays are not empty. This means that if deck doesn't have cards of manaValue '2', for example, the corresponding div will not be there, and the next (not empty) div will occupy it's space.